improved unit test for history

+++ src/main/java/edu/aust/order/dto/OrderRequest.java
@@ -5,7 +5,7 @@ public record OrderRequest(
         long productId,
         String productName,
         int quantity,
-        double price,
+        float price,
         String status
 ) {
