+++ src/main/java/edu/aust/order/utilities/MathUtil.java
@@ -2,8 +2,8 @@ package edu.aust.order.utilities;
 
 public class MathUtil {
 
-    public static final double getTotalPrice(int quantity, double pricePerUnit) {
-        return ((double) quantity) * pricePerUnit;
+    public static final float getTotalPrice(int quantity, float pricePerUnit) {
+        return quantity * pricePerUnit;
     }
 
 }
